feat(AI.py): Update AI Class
